:recycle: refactor: log 의 모델 수정 및 createHabitLog 위치 수정
- log 클래스에서 habitTypeId 추가
- mapping 시 id를 잘못 전달하는 오류 수정
- createHabitLog 를 data 패키지로 옮기기

Issue: #41
